Add PushContainer, move device api to PushContainer
